website/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-allauth==0.42.0
django-extensions==3.0.7
django-model-utils==4.0.0
django-notifications-hq==1.6.0
djangorestframework==3.11.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jsonfield==3.1.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.8
webencodings==0.5.1
